requirements.txt
requests
Pillow
schedule
playsound
